Move lots of functionality under lib.rs
