+++ src/products.rs
@@ -2,6 +2,8 @@
 #![forbid(unsafe_code)]
 #![forbid(missing_docs)]
 
+/// A list of [HashiCorp](https://www.hashicorp.com/) products that this crate
+/// can download.
 pub const PRODUCTS_LIST: &[&str] = &[
     "consul",
     "nomad",
